requirements.txt
# Django e pacotes relacionados
Django==4.2.7
django-crispy-forms==2.0
crispy-bootstrap4==2022.1
django-ckeditor==6.7.0
django-cleanup==8.0.0

# Processamento de imagens
Pillow==10.1.0

# Processamento de datas
python-dateutil==2.8.2

# Pacotes de segurança
django-cors-headers==4.3.0
django-axes==6.1.1

# Ferramentas de desenvolvimento
django-debug-toolbar==4.2.0

# Produção
gunicorn==21.2.0
whitenoise==6.6.0

# Para testes
pytest==7.4.3
pytest-django==4.7.0

# Banco de dados (descomente para produção)
# psycopg2-binary==2.9.9  # Para PostgreSQL